models, repositories, controllers, dtos, services aangemaakt voor remote controller, wall bracket en ci module
