+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/controllers/RemoteControlController.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.controllers;
+
+public class RemoteControlController {
+}
