+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/dtos/ciModules/CiModuleDto.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.dtos.ciModules;
+
+public class CiModuleDto {
+}
