+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/dtos/ciModules/CiModuleInputDto.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.dtos.ciModules;
+
+public class CiModuleInputDto {
+}
