+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/dtos/remoteControls/RemoteControlInputDto.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.dtos.remoteControls;
+
+public class RemoteControlInputDto {
+}
