+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/dtos/wallBrackets/WallBracketInputDto.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.dtos.wallBrackets;
+
+public class WallBracketInputDto {
+}
