+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/models/CiModule.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.models;
+
+public class CiModule {
+}
