+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/models/RemoteControl.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.models;
+
+public class RemoteControl {
+}
