+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/repositories/CiModuleRepository.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.repositories;
+
+public interface CiModuleRepository {
+}
