+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/repositories/WallBracketRepository.java
@@ -1,0 +1,4 @@
+package com.example.backendspringboottechiteasycontrollerles10.repositories;
+
+public interface WallBracketRepository {
+}
