+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/services/CiModuleService.java
@@ -1,0 +1,2 @@
+package com.example.backendspringboottechiteasycontrollerles10.services;public class CiModuleService {
+}
