+++ src/main/java/com/example/backendspringboottechiteasycontrollerles10/services/RemoteControlService.java
@@ -1,0 +1,2 @@
+package com.example.backendspringboottechiteasycontrollerles10.services;public class RemoteControlService {
+}
