--- src/main/resources/data.sql
@@ -1,2 +1,0 @@
-INSERT INTO televisions
-VALUES
